+++ lib/deimos/schema_class/enum.rb
@@ -39,7 +39,7 @@ module Deimos
       end
 
       # @return [SchemaClass::Enum]
-      def self.initialize_from_value(value)
+      def self.initialize_from_value(value, from_message: false)
         return nil if value.nil?
 
         value.is_a?(self) ? value : self.new(value)