+++ lib/deimos/schema_class/record.rb
@@ -72,11 +72,30 @@ module Deimos
         validator.schema_fields.map(&:name)
       end
 
+      # Used internally within Deimos so that we don't crash on unknown fields that come from
+      # a backwards compatible schema.
+      # @param kwargs [Hash] the attributes to set on the new object.
       # @return [SchemaClass::Record]
-      def self.initialize_from_value(value)
+      def self.new_from_message(**kwargs)
+        @from_message = true
+        record = self.new
+        kwargs.each do |k, v|
+          record.send("#{k}=", v) if record.respond_to?("#{k}=")
+        end
+        record
+      end
+
+      # @return [SchemaClass::Record]
+      # @param from_message [Boolean] whether it's being initialized from a real Avro message.
+      def self.initialize_from_value(value, from_message: false)
         return nil if value.nil?
 
-        value.is_a?(self) ? value : self.new(**value.symbolize_keys)
+        return value if value.is_a?(self)
+        if from_message
+          self.new(**value.symbolize_keys)
+        else
+          self.new_from_message(**value.symbolize_keys)
+        end
       end
     end
   end